1. Thêm hàm autoCloseAuction - ManagerService 2. Xử lý lock - placeBid - BiddingService 3.handleException bằng package exception để định nghĩa từng lỗi 4. Logic Admin 5. Giáo diện Admin về RAM

+++ src/main/java/com/auction/common/model/Auction.java
@@ -51,7 +51,7 @@ public class Auction {
     }
 
     //cập nhật giá khi có người trả giá cao hơn
-    public void setCurrentPrice(int currentPrice) {
+    public void setCurrentPrice(long currentPrice) {
         this.currentPrice = currentPrice;
     }
